Split the CUDA Gravity routine into its own .cu file
